@@ -10,19 +10,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.0",
-    "@angular/common": "~13.0.0",
-    "@angular/compiler": "~13.0.0",
-    "@angular/core": "~13.0.0",
-    "@angular/forms": "~13.0.0",
-    "@angular/platform-browser": "~13.0.0",
-    "@angular/platform-browser-dynamic": "~13.0.0",
-    "@angular/router": "~13.0.0",
+    "@abacritt/angularx-social-login": "^1.1.0",
+    "@angular/animations": "~14.0.2",
+    "@angular/common": "~14.0.2",
+    "@angular/compiler": "~14.0.2",
+    "@angular/core": "~14.0.2",
+    "@angular/forms": "~14.0.2",
+    "@angular/platform-browser": "~14.0.2",
+    "@angular/platform-browser-dynamic": "~14.0.2",
+    "@angular/router": "~14.0.2",
     "@fortawesome/angular-fontawesome": "^0.10.2",
     "@fortawesome/fontawesome-free": "^6.1.1",
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
     "@popperjs/core": "^2.11.5",
+    "angularx-social-login": "^4.1.0",
     "bootstrap": "^5.1.3",
     "install": "^0.13.0",
     "jquery": "^3.6.0",
@@ -33,9 +35,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.0.3",
-    "@angular/cli": "~13.0.3",
-    "@angular/compiler-cli": "~13.0.0",
+    "@angular-devkit/build-angular": "^14.0.2",
+    "@angular/cli": "^14.0.2",
+    "@angular/compiler-cli": "~14.0.2",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~3.10.0",
@@ -44,6 +46,6 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.4.3"
+    "typescript": "~4.7.3"
   }
 }
