@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import { AuthenticationService } from 'src/app/services/security/authentication.service';
 import { SpaceValidator } from '../space-validator';
@@ -11,11 +11,11 @@ import { SpaceValidator } from '../space-validator';
 })
 export class CodeActivationComponent implements OnInit {
 
-  formParentGroup!: FormGroup;
+  formParentGroup!: UntypedFormGroup;
 
   email: string ="";
 
-  constructor(private formChildGroup: FormBuilder,private authenticationService: AuthenticationService, private router:Router) { }
+  constructor(private formChildGroup: UntypedFormBuilder,private authenticationService: AuthenticationService, private router:Router) { }
 
   ngOnInit(): void {
     this.email = sessionStorage.getItem("emailActive");
@@ -26,7 +26,7 @@ export class CodeActivationComponent implements OnInit {
   myLoginForm(){
     this.formParentGroup = this.formChildGroup.group({
       user: this.formChildGroup.group({
-        code: new FormControl('',[Validators.required,
+        code: new UntypedFormControl('',[Validators.required,
                                   SpaceValidator.onlyContainSpace])
       })
     })
