@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
+import { GoogleLoginProvider, FacebookLoginProvider, SocialAuthService } from 'angularx-social-login';
 import { AuthenticationService } from 'src/app/services/security/authentication.service';
 import { SpaceValidator } from '../space-validator';
 
@@ -11,12 +12,13 @@ import { SpaceValidator } from '../space-validator';
 })
 export class LoginComponent implements OnInit {
 
-  formParentGroup : FormGroup;
+  formParentGroup : UntypedFormGroup;
   submitted = false;
 
-  constructor(private formChildGroup: FormBuilder,
+  constructor(private formChildGroup: UntypedFormBuilder,
               private authenticationService: AuthenticationService,
-              private router: Router) { }
+              private router: Router,
+              private authService: SocialAuthService) { }
 
   ngOnInit(): void {
     this.myLoginForm();
@@ -25,10 +27,10 @@ export class LoginComponent implements OnInit {
   myLoginForm(){
     this.formParentGroup = this.formChildGroup.group({
       user: this.formChildGroup.group({
-        email: new FormControl('',[Validators.required,
+        email: new UntypedFormControl('',[Validators.required,
                                   SpaceValidator.onlyContainSpace,
                                   Validators.pattern('^[\\w-\\.]+@([\\w-]+\\.)+[\\w-]{2,4}$') ]),
-        password: new FormControl('', [Validators.required])
+        password: new UntypedFormControl('', [Validators.required])
       })
     })
   }
@@ -73,8 +75,27 @@ login(){
           }
         })
 
-
       }
+
+
+
+      signInWithGoogle(): void {
+        this.authService.signIn(GoogleLoginProvider.PROVIDER_ID).then(
+          data => {
+            console.log(data)
+          }
+        );
+      }
+
+      signInWithFB(): void {
+        this.authService.signIn(FacebookLoginProvider.PROVIDER_ID).then(
+          data =>{
+            console.log(data)
+          }
+        );
+      }
+
+
 
 
 
