@@ -1,8 +1,8 @@
-import { FormControl, ValidationErrors } from '@angular/forms';
+import { UntypedFormControl, ValidationErrors } from '@angular/forms';
 
 export class SpaceValidator {
 
-  static onlyContainSpace(control: FormControl): ValidationErrors | null{
+  static onlyContainSpace(control: UntypedFormControl): ValidationErrors | null{
     if(control.value !=null && control.value.trim().length === 0){
       return {'noSpace': true}
     }
